TEXTOIR_ct_adb/open_intent_detection/requirements.txt
boto3==1.18.13
botocore==1.21.13
certifi==2021.5.30
charset-normalizer==2.0.4
easydict>=1.9
idna==3.2
jmespath==0.10.0
joblib==1.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
pandas==1.1.5
Pillow>=8.2.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
threadpoolctl==2.2.0
transformers==4.16.2
torchvision>=0.8.2
tqdm==4.62.0
urllib3==1.26.6